requirements.txt
numpy
pandas
matplotlib
pillow
tqdm
scikit-learn
opencv-python
xgboost
torch
torchvision
mediapipe